third_party/WebKit/LayoutTests/platform/mac-retina/virtual/rootlayerscrolls/fast/scrolling/fixed-position-behind-scrollbar-expected.txt

FAIL scrollY > 0 failed to change to true in 0.5 seconds.
PASS 205,205 is 205,205
PASS successfullyParsed is true

TEST COMPLETE
Tests that the scrollbar can be clicked even when it clips a position:fixed element, and that the scrollbars are excluded from the viewport for determining the element's position.
